LPS-101645 Deprecate `Liferay.Util.focusFormField`
The goal of this change is to deprecate the `Liferay.Util.focusFormField`
function: it will now be available as a module (i.e. it can be `import`ed),
as well as a global to avoid breaking changes.

This function was also using the following functions:
- `getDOM`
- `getElement`
- `inBrowserView`
that have also been moved to the `frontend-js-web` module.

An easy way to test this, after having fetched these changes, it to
deploy the `frontend-js-aui-web` and `frontend-js-web` modules, and to
sign into liferay-portal (or sign out  and sign in, if already signed
in).
