+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/util/get_element.js
@@ -1,0 +1,25 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import getDOM from './get_dom';
+
+export default function getElement(el) {
+	const currentElement = getDOM(el);
+
+	return typeof currentElement === 'string'
+		? document.querySelector(currentElement)
+		: currentElement.jquery
+		? currentElement[0]
+		: currentElement;
+}
